[Entity Store] Move history_snapshot_client to history_snapshot
