+++ x-pack/solutions/security/plugins/entity_store/server/domain/history_snapshot/constants.ts
