+++ x-pack/solutions/security/plugins/entity_store/server/request_context_factory.ts
@@ -20,7 +20,7 @@ import {
   EntityStoreGlobalStateClient,
 } from './domain/definitions/saved_objects';
 import { CcsLogsExtractionClient, LogsExtractionClient } from './domain/logs_extraction';
-import { HistorySnapshotClient } from './domain/history_snapshot_client';
+import { HistorySnapshotClient } from './domain/history_snapshot';
 import { CRUDClient } from './domain/crud_client';
 import type { TelemetryReporter } from './telemetry/events';
