+++ x-pack/solutions/security/plugins/entity_store/server/tasks/history_snapshot_task.ts
@@ -17,7 +17,7 @@ import { TasksConfig } from './config';
 import { EntityStoreTaskType } from './constants';
 import type { EntityStoreCoreSetup } from '../types';
 import { EntityStoreGlobalStateClient } from '../domain/definitions/saved_objects';
-import { HistorySnapshotClient } from '../domain/history_snapshot_client';
+import { HistorySnapshotClient } from '../domain/history_snapshot';
 
 const config = TasksConfig[EntityStoreTaskType.enum.historySnapshot];
