+++ x-pack/solutions/security/plugins/entity_store/server/types.ts
@@ -28,7 +28,7 @@ import type { AssetManager } from './domain/asset_manager';
 import type { EntityMaintainersClient } from './domain/entity_maintainers';
 import type { FeatureFlags } from './infra/feature_flags';
 import type { CcsLogsExtractionClient, LogsExtractionClient } from './domain/logs_extraction';
-import type { HistorySnapshotClient } from './domain/history_snapshot_client';
+import type { HistorySnapshotClient } from './domain/history_snapshot';
 import type { CRUDClient } from './domain/crud_client';
 import type { RegisterEntityMaintainerConfig } from './tasks/entity_maintainers/types';
